WIP Details screen

+++ domain/src/main/java/com/weatherapp/domain/model/Place.kt
@@ -1,6 +1,6 @@
 package com.weatherapp.domain.model
 
 sealed interface Place {
-    data class Coordinates(val latitude: Float, val longitude: Float) : Place
+    data class Coordinates(val latitude: Double, val longitude: Double) : Place
     data class Address(val address: String) : Place
 }
